Create RansomNote.java
You will get the question details here ;- https://www.hackerrank.com/challenges/ctci-ransom-note/problem?h_l=interview&playlist_slugs%5B%5D=interview-preparation-kit&playlist_slugs%5B%5D=dictionaries-hashmaps

Appraoch 👍 
step 1 :- Count frequencey of all the word in magazine and put it in HashMap
step2 :- for each word of note search in hashMap whether it is present or not. If yes the subtract -1 and update it
step 3 :- if key is not present in hashmap or if present its value is 0 that means note is not valid. Reture No
step 4:- if step 3 is not correct that means note is valid. return Yes